+++ Classes/Exception/VariableResolverException.php
@@ -9,6 +9,6 @@ declare(strict_types=1);
  * LICENSE.txt file that was distributed with this source code.
  */
 
-namespace Brotkrueml\JobRouterBase\Exception;
+namespace JobRouter\AddOn\Typo3Base\Exception;
 
 final class VariableResolverException extends \RuntimeException {}