+++ Classes/Widgets/Provider/TransferStatusDataProviderInterface.php
@@ -9,9 +9,9 @@ declare(strict_types=1);
  * LICENSE.txt file that was distributed with this source code.
  */
 
-namespace Brotkrueml\JobRouterBase\Widgets\Provider;
+namespace JobRouter\AddOn\Typo3Base\Widgets\Provider;
 
-use Brotkrueml\JobRouterBase\Domain\Dto\TransferStatus;
+use JobRouter\AddOn\Typo3Base\Domain\Dto\TransferStatus;
 
 /**
  * @internal Only for usage in the TYPO3 JobRouter extensions!