+++ Configuration/Icons.php
@@ -12,14 +12,14 @@ declare(strict_types=1);
 return [
     'jobrouter-base-status-failed' => [
         'provider' => TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
-        'source' => 'EXT:' . Brotkrueml\JobRouterBase\Extension::KEY . '/Resources/Public/Icons/status-failed.svg',
+        'source' => 'EXT:' . JobRouter\AddOn\Typo3Base\Extension::KEY . '/Resources/Public/Icons/status-failed.svg',
     ],
     'jobrouter-base-status-pending' => [
         'provider' => TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
-        'source' => 'EXT:' . Brotkrueml\JobRouterBase\Extension::KEY . '/Resources/Public/Icons/status-pending.svg',
+        'source' => 'EXT:' . JobRouter\AddOn\Typo3Base\Extension::KEY . '/Resources/Public/Icons/status-pending.svg',
     ],
     'jobrouter-base-status-successful' => [
         'provider' => TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
-        'source' => 'EXT:' . Brotkrueml\JobRouterBase\Extension::KEY . '/Resources/Public/Icons/status-successful.svg',
+        'source' => 'EXT:' . JobRouter\AddOn\Typo3Base\Extension::KEY . '/Resources/Public/Icons/status-successful.svg',
     ],
 ];