+++ Documentation/changelog.rst
@@ -11,6 +11,12 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 `Unreleased <https://github.com/brotkrueml/typo3-jobrouter-base/compare/v2.0.0...HEAD>`_
 --------------------------------------------------------------------------------------------
 
+Changed
+^^^^^^^
+
+
+* Namespace from ``Brotkrueml\JobRouterBase`` to ``JobRouter\Addon\Typo3Base``
+
 `2.0.0 <https://github.com/brotkrueml/typo3-jobrouter-base/compare/v1.3.0...v2.0.0>`_ - 2023-05-31
 ------------------------------------------------------------------------------------------------------